busca_semantica_pgvector/requirements.txt
psycopg[binary]
sentence-transformers
python-dotenv
uvicorn
fastapi
pydantic
pgvector